Fix whitespace formatting issues
Used autopep to fix only the most common whitespace issues. In 
particular, lines that were too long were not fixed. The exact command 
was:
```
autopep8 tools/RAiDER/ --recursive --select 
E231,E225,E251,E111,E128,E127,E265,E303,E261,E301 --in-place
```

+++ tools/RAiDER/models/generateGACOSVRT.py
@@ -4,14 +4,14 @@
 # Copyright 2018
 
 
-def makeVRT(filename, dtype = 'Float32'):
+def makeVRT(filename, dtype='Float32'):
     '''
     Use an RSC file to create a GDAL-compatible VRT file for opening GACOS weather model files
     '''
     fields = readRSC(filename)
-    string = vrtStr(fields['XMAX'],fields['YMAX'], fields['X_FIRST'],fields['Y_FIRST'], fields['X_STEP'],fields['Y_STEP'],filename.replace('.rsc', ''), dtype = dtype)
+    string = vrtStr(fields['XMAX'], fields['YMAX'], fields['X_FIRST'], fields['Y_FIRST'], fields['X_STEP'], fields['Y_STEP'], filename.replace('.rsc', ''), dtype=dtype)
     writeStringToFile(string, filename.replace('.rsc', '').replace('.ztd', '') + '.vrt')
-        
+
 
 def writeStringToFile(string, filename):
     '''
@@ -19,7 +19,7 @@ def writeStringToFile(string, filename):
     '''
     with open(filename, 'w') as f:
         f.write(string)
-    
+
 
 def readRSC(rscFilename):
     fields = {}
@@ -28,9 +28,9 @@ def readRSC(rscFilename):
             fieldName, value = line.strip().split()
             fields[fieldName] = value
     return fields
-        
 
-def vrtStr(xSize, ySize, lon1, lat1, lonStep, latStep, filename, dtype = 'Float32'):
+
+def vrtStr(xSize, ySize, lon1, lat1, lonStep, latStep, filename, dtype='Float32'):
     string = '''<VRTDataset rasterXSize="{xSize}" rasterYSize="{ySize}">
   <SRS>EPSG:4326</SRS>
   <GeoTransform> {lon1}, {lonStep},  0.0000000000000000e+00,  {lat1},  0.0000000000000000e+00, {latStep}</GeoTransform>
@@ -38,7 +38,7 @@ def vrtStr(xSize, ySize, lon1, lat1, lonStep, latStep, filename, dtype = 'Float3
     <SourceFilename relativeToVRT="1">{filename}</SourceFilename>
   </VRTRasterBand>
 </VRTDataset>
-'''.format(xSize=xSize, ySize=ySize, filename=filename, dtype=dtype, lon1 = lon1, lat1 = lat1)
+'''.format(xSize=xSize, ySize=ySize, filename=filename, dtype=dtype, lon1=lon1, lat1=lat1)
 
     return string
 
@@ -53,14 +53,14 @@ def convertAllFiles(dirLoc):
         makeVRT(f)
 
 
-if __name__=='__main__': 
-   import sys
-   if len(sys.argv) == 2:
-      makeVRT(sys.argv[1])
-   elif len(sys.argv) == 3:
-      convertAllFiles(sys.argv[1])
-      print('Converting all RSC files in {}'.format(sys.argv[1]))
-   else:
-      print('Usage: ')
-      print('python3 generateGACOSVRT.py <rsc_filename>')
-      sys.exit(0)
+if __name__ == '__main__': 
+    import sys
+    if len(sys.argv) == 2:
+        makeVRT(sys.argv[1])
+    elif len(sys.argv) == 3:
+        convertAllFiles(sys.argv[1])
+        print('Converting all RSC files in {}'.format(sys.argv[1]))
+    else:
+        print('Usage: ')
+        print('python3 generateGACOSVRT.py <rsc_filename>')
+        sys.exit(0)
